saving uploaded files

+++ internal/entities/file.go
@@ -9,12 +9,12 @@ import (
 type DbStoreFileParams struct {
 	ID			uuid.UUID
 	Filename 	string
-	Filesize 	int32
+	Filesize 	int64
 }
 
 type File struct {
 	ID            uuid.UUID  `json:"id"`
 	Filename      string     `json:"filename"`
-	Filesize      int32      `json:"filesize"`
+	Filesize      int64      `json:"filesize"`
 	UploadedAt	  time.Time	 `json:"uploaded_at"`
 }
